Refactor: Move RehauNeaSmartHeatarea to separate module
