+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+- Refactored project structure: moved `RehauNeaSmartHeatarea` to `pyrehau_neasmart/heatarea.py`.
+- Extracted constants to `const.py`, exceptions to `exceptions.py`, and utils to `utils.py`.
+
 ## [0.0.7] - 2026-01-21
 
 ### Changed
