+++ pyrehau_neasmart/const.py
@@ -1,0 +1,9 @@
+HEATAREA_MODES = {
+    0: "AUTO",
+    1: "COMFORT",
+    2: "ECO",
+}
+
+_MODES = ["auto", "day", "night"]
+_PATH_GET = '/data/static.xml'
+_PATH_POST = '/data/changes.xml'
